fix v1, add tracers and printfs next times
this might be too much out of my scope, but what I think happens, its by error, it delets the 2 compents itself or doesn't like the code with double same duplicate components, and just freaks from the natural lnauge code. or I need to be more carefull in editor, and save mroe.